refactor: 예외 클래스 네이밍 변경 - #14 - ForbiddenException 이 보다 명확하다고 판단함
