+++ src/main/java/org/example/capstonedesign1/global/exception/ForbiddenException.java
@@ -2,12 +2,12 @@ package org.example.capstonedesign1.global.exception;
 
 import org.example.capstonedesign1.global.exception.code.ErrorCode;
 
-public class AuthorizedException extends CustomException{
-    public AuthorizedException(ErrorCode errorCode) {
+public class ForbiddenException extends CustomException {
+    public ForbiddenException(ErrorCode errorCode) {
         super(errorCode);
     }
 
-    public AuthorizedException(ErrorCode errorCode, String detail) {
+    public ForbiddenException(ErrorCode errorCode, String detail) {
         super(errorCode, detail);
     }
 }
